Core Check-in - First Check-in
A refactoring of the AutoTaskTicketManagerWorkerService-L. (The L was for Linux in the cross platform version)
In this checking I have the Basic Configuration, Authentication, and MSGraph MSCAL processes running. Basically that means that the app core is written and can get a Azure Authentication Bearer Token and the root of the Internal Management API is active.